+++ src/main/java/com/redhat/devtools/intellij/tektoncd/ui/wizard/OutputResourcesStep.java
@@ -13,7 +13,6 @@ package com.redhat.devtools.intellij.tektoncd.ui.wizard;
 import com.redhat.devtools.intellij.tektoncd.tkn.Resource;
 import com.redhat.devtools.intellij.tektoncd.tkn.component.field.Output;
 import com.redhat.devtools.intellij.tektoncd.utils.model.actions.ActionToRunModel;
-import com.redhat.devtools.intellij.tektoncd.utils.model.actions.StartResourceModel;
 import java.awt.GridBagConstraints;
 import javax.swing.JComboBox;
 import javax.swing.JLabel;
@@ -39,10 +38,10 @@ public class OutputResourcesStep extends BaseStep {
         return "https://github.com/tektoncd/pipeline/blob/master/docs/resources.md";
     }
 
-    public void setContent(ActionToRunModel model) {
+    public void setContent() {
         final int[] row = {0};
 
-        model.getOutputResources().values().stream().forEach(output -> {
+        model.getOutputResources().stream().forEach(output -> {
             JLabel lblNameResource = new JLabel("<html><span style=\\\"font-family:serif;font-size:10px;font-weight:bold;\\\">" + output.name() + "</span></html");
             addComponent(lblNameResource, null, BORDER_LABEL_NAME, ROW_DIMENSION, 0, row[0], GridBagConstraints.NORTH);
             addTooltip(lblNameResource, output.description().orElse(""));
@@ -80,7 +79,7 @@ public class OutputResourcesStep extends BaseStep {
     }
 
     private void setOutputValue(String outputName, String value) {
-        for (Output output : model.getOutputResources().values()) {
+        for (Output output : model.getOutputResources()) {
             if (output.name().equals(outputName)) {
                 output.setValue(value);
                 break;
