+++ src/main/java/com/redhat/devtools/intellij/tektoncd/utils/model/ConfigurationModelFactory.java
@@ -43,22 +43,4 @@ public class ConfigurationModelFactory {
 
         return null;
     }
-
-    public static ResourceConfigurationModel getResourceModel(String configuration) {
-        try {
-            String kind = YAMLHelper.getStringValueFromYAML(configuration, new String[] {"kind"});
-            if (!Strings.isNullOrEmpty(kind)) {
-                switch (kind.toLowerCase()) {
-                    case "pipeline": return new PipelineConfigurationModel(configuration);
-                    case "task": return new TaskConfigurationModel(configuration);
-                    case "condition": return new ConditionConfigurationModel(configuration);
-                    default: break;
-                }
-            }
-        } catch (IOException e) {
-            logger.warn(e.getLocalizedMessage());
-        }
-
-        return null;
-    }
 }
